% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mturk_workers}
\alias{mturk_workers}
\title{Data from mmbruteluck - Online Experiment, mTurk workers}
\format{
A data frame with 2437 observations and 17 variables:
\describe{
\item{id}{Non-informative identifier of individual participant}
\item{female}{Indicator for mTurker being female (0/1)}
\item{age}{Age of mTurker in years (numeric) }
\item{education}{Education category
\describe{
 \item{1.}{Less than High School}
 \item{2.}{High School/GED  }
 \item{3.}{Some College  }
 \item{4.}{2 year College Degree }
 \item{5.}{4 year College Degree }
 \item{6.}{Masters Degree }
 \item{7.}{Masters Degree  }
 \item{8.}{Professional Degree (JD, MD)}
}
}
\item{political}{Political identification.Numeric, [-2, - 2],
with -2 being "very liberal" and 2 "very conservative.}
\item{treatment}{Treatment arm (characters)}
\item{safe}{Did the mTurker choose the safe alternative? (0/1), 1 if they
chose the safe alternative}
\item{bonus_ingame}{bonus in game (0.5USD) - 0.25 USD if worker chose
costly safe alternative in that treatment}
\item{bonus_lottery}{Outcome of lottery choice prior to redistribution (numeric, in USD)}
\item{bonus_final}{The sum of bonus_ingame and the post redistribution income from lottery (numeric, in USD)}
\item{repeater}{Has the worker attempted to start the survey more than once?
(0/1), 1 if they did so.}
\item{lottery_or_green}{Indicator for whether the worker chose the lottery in Forced Choice treatments or the
green ball in Nominal Choice treatments (0/1), 1 if they dis so.}
\item{duration_survey}{Duration of HIT completion (numeric, in seconds)}
\item{treatment_str}{Treatment (characters)}
\item{bonus_startup}{Bonus account at startup (numeric, 0.5 USD)}
\item{winner}{Indicator for whether the worker is a lottery winner (0/1).}
\item{loser}{Indicator for whether the worker is a lottery loser (0/1).}
}
}
\source{
Experiment conducted by the authors of the paper.
}
\usage{
mturk_workers
}
\description{
A dataset from the online experiment in the paper
"Choice and personal responsibility: What is a
morally relevant choice?" by Alexander W. Cappelen, Sebastian
Fest, Erik Ø. Sørensen, and Bertil Tungodden.
}
\keyword{datasets}
